Add shared `logs.py` module, add docs link
