+++ mreg/middleware/logging_http.py
@@ -2,14 +2,16 @@
 import http
 import logging
 import time
+import traceback
 import uuid
 from typing import Callable, cast
 
-import structlog
 import sentry_sdk
-import traceback
+import structlog
 from django.conf import settings
 from django.http import HttpRequest, HttpResponse
+
+from mreg.logs import get_request_body, get_request_header, get_request_user_agent, get_request_username
 
 mreg_logger = structlog.getLogger("mreg.http")
 
@@ -20,47 +22,6 @@ LOGMAP = {
     "INFO": logging.INFO,
     "DEBUG": logging.DEBUG,
 }
-
-def get_request_header(
-    request: HttpRequest, header_key: str, meta_key: str
-) -> str:
-    """Get the value of a header from the request, either via headers or META."""
-    if hasattr(request, "headers"): # pragma: no cover
-        return request.headers.get(header_key)
-    return request.META.get(meta_key)
-
-
-def get_request_body(request: HttpRequest) -> str:
-    """Get the request body as a string, or '<Binary Data>' if it's binary.
-
-    We currently do not support multipart/form-data requests.
-    """
-    if request.POST:
-        return request.POST.dict()
-
-    try:
-        body = request.body.decode("utf-8")
-    except UnicodeDecodeError:
-        return "<Binary Data>"
-
-    # Try to remove the content-type line and leading line breaks
-    body = body.split("\n", 1)[-1]  # Removes the first line
-    body = body.lstrip()  # Removes leading line breaks
-
-    # Limit the size of the body logged
-    return body[: settings.LOGGING_MAX_BODY_LENGTH]
-
-
-def get_request_username(request: HttpRequest, default: str = "AnonymousUser") -> str:
-    """Get the username of the user making the request, or 'AnonymousUser'."""
-    return getattr(request.user, "username", default) if hasattr(request, 'user') else default
-
-
-def get_request_user_agent(request: HttpRequest, default: str = "Unknown") -> str:
-    try:
-        return get_request_header(request, "user-agent", "HTTP_USER_AGENT")
-    except Exception:
-        return default
 
 
 class LoggingMiddleware:
